pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "community-briefing"
version = "0.1.0"
description = "Multi-community topic briefing API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "apscheduler==3.11.0",
  "beautifulsoup4==4.13.4",
  "fastapi==0.116.1",
  "httpx==0.28.1",
  "pydantic-settings==2.11.0",
  "psycopg[binary]==3.2.10",
  "python-dateutil==2.9.0.post0",
  "sqlalchemy==2.0.43",
  "uvicorn==0.35.0",
  "scikit-learn==1.7.1",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
  "pytest-cov==6.2.1",
  "ruff==0.12.11",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]
